pyproject.toml
[project]
name = "refyn-ai"
version = "0.1.0"
description = "The code reviewer that learns your team"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic-settings>=2.2.1",
    "qdrant-client>=1.8.0",
    "redis>=5.0.3",
    "asyncpg>=0.29.0",
    "tree-sitter>=0.21.3",
    "tree-sitter-python>=0.21.0",
    "langgraph>=0.0.31",
    "langchain-openai>=0.1.1",
    "langchain-core>=0.1.33",
    "langchain>=0.1.12",
    "structlog>=24.1.0",
    "httpx>=0.27.0",
    "alembic>=1.13.1",
    "apscheduler>=3.10.4",
    "pyjwt[crypto]>=2.8.0",
    "cryptography>=42.0.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "ruff>=0.3.4",
    "mypy>=1.9.0",
    "respx>=0.21.0",
    "testcontainers[postgres,redis]>=3.7.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "rag", "agents", "feedback", "dashboard"]

[tool.ruff]
line-length = 100

[tool.mypy]
strict = true
